refactor(data): relocate query I/O and countries to canonical layers
- move data/supabase/queries.ts + queries/core.ts to src/lib/supabase/{queries,
  queries-core}.ts (the single Supabase-I/O home); data/supabase/queries.ts is now a
  re-export shim so all ~44 @/data/supabase/queries importers keep working unchanged
- relocate static lib/data/countries.ts to src/data/countries.ts (static belongs in
  src/data); repoint its one importer
- drops src/data boundary warnings 66 -> 21 (only the guide-assets/guide-templates
  browser modules remain); ratchet floor lowered to 21
